Split out setup to its own sub project
