+++ build.sbt
@@ -169,3 +169,11 @@ lazy val configTools = (project in file("configTools"))
     name := "janus-config-tools",
     description := "Library for reading and writing Janus configuration files"
   )
+
+lazy val Setup = (project in file("setup"))
+  .dependsOn(root)
+  .settings(
+    commonSettings,
+    name := "janus-devcontainer-setup",
+    description := "Library for setting up dynamodb tables"
+  )
