+++ scripts/setup.sh
@@ -35,20 +35,7 @@ docker compose -f local-dev/docker-compose.yml up -d
 sleep 2
 
 # Creating tables (if needed - aim for idempotency)
-aws dynamodb list-tables --profile security --region eu-west-1 --endpoint http://localhost:8000 \
-  | jq -r '.TableNames[]' \
-  | grep -w PasskeyChallenges  \
-  || sbt "Test / runMain aws.PasskeyChallengeDBTestApp create"
-
-aws dynamodb list-tables --profile security --region eu-west-1 --endpoint http://localhost:8000 \
-  | jq -r '.TableNames[]' \
-  | grep -w Passkeys \
-  || sbt "Test / runMain aws.PasskeyDBTestApp create"
-
-aws dynamodb list-tables --profile security --region eu-west-1 --endpoint http://localhost:8000 \
-  | jq -r '.TableNames[]' \
-  | grep -w AuditTrail \
-  || sbt "Test / runMain aws.AuditTrailDBTestApp create"
+sbt "Setup / runMain aws.DBSetupApp"
 
 if [[ ! -f ~/.gu/janus-app/janusData.conf ]]; then
    echo "!!! You need to copy in a janusData.conf file at ~/.gu/janus-app/janusData.conf"
