+++ setup/src/main/scala/aws/PasskeyDBSetup.scala
@@ -1,0 +1,63 @@
+package aws
+
+import software.amazon.awssdk.services.dynamodb.DynamoDbClient
+import software.amazon.awssdk.services.dynamodb.model.KeyType.{HASH, RANGE}
+import software.amazon.awssdk.services.dynamodb.model.ScalarAttributeType.S
+import software.amazon.awssdk.services.dynamodb.model._
+
+class PasskeyDBSetup {
+
+  /** NB: Only use these for local testing use the provided CloudFormation
+    * template to create table in AWS environments.
+    *
+    * If you update this then be sure to also update the CloudFormation
+    * template's definition.
+    */
+  private[aws] def createTable()(implicit
+      dynamoDB: DynamoDbClient
+  ): CreateTableResponse =
+    dynamoDB.createTable(
+      CreateTableRequest
+        .builder()
+        .tableName(PasskeyDB.tableName)
+        .keySchema(
+          KeySchemaElement
+            .builder()
+            .attributeName("username")
+            .keyType(HASH)
+            .build(),
+          KeySchemaElement
+            .builder()
+            .attributeName("credentialId")
+            .keyType(RANGE)
+            .build()
+        )
+        .attributeDefinitions(
+          AttributeDefinition
+            .builder()
+            .attributeName("username")
+            .attributeType(S)
+            .build(),
+          AttributeDefinition
+            .builder()
+            .attributeName("credentialId")
+            .attributeType(S)
+            .build()
+        )
+        .provisionedThroughput(
+          ProvisionedThroughput
+            .builder()
+            .readCapacityUnits(15L)
+            .writeCapacityUnits(15L)
+            .build()
+        )
+        .build()
+    )
+
+  private[aws] def destroyTable()(implicit
+      dynamoDb: DynamoDbClient
+  ): DeleteTableResponse =
+    dynamoDb.deleteTable(
+      DeleteTableRequest.builder().tableName(PasskeyDB.tableName).build()
+    )
+}
