+++ test/aws/AuditTrailDBTest.scala
@@ -45,113 +45,6 @@ class AuditTrailDBTest extends AnyFreeSpec with Matchers {
       println(userResults.toList)
     }
 
-    "create table" ignore {
-      createTable()
-    }
-
-    "destroy table" ignore {
-      destroyTable()
-    }
-  }
-
-  /** NB: Only use these for local testing use the provided CloudFormation
-    * template to create table in AWS environments.
-    *
-    * If you update this then be sure to also update the CloudFormation
-    * template's definition.
-    */
-  private[aws] def createTable()(implicit
-      dynamoDB: DynamoDbClient
-  ): CreateTableResponse = {
-    val auditTableCreateRequest = CreateTableRequest
-      .builder()
-      .tableName(AuditTrail.tableName)
-      .keySchema(
-        KeySchemaElement
-          .builder()
-          .attributeName(AuditTrail.accountPartitionKeyName)
-          .keyType(HASH)
-          .build(),
-        KeySchemaElement
-          .builder()
-          .attributeName(AuditTrail.timestampSortKeyName)
-          .keyType(RANGE)
-          .build()
-      )
-      .attributeDefinitions(
-        AttributeDefinition
-          .builder()
-          .attributeName(AuditTrail.accountPartitionKeyName)
-          .attributeType(S)
-          .build(),
-        AttributeDefinition
-          .builder()
-          .attributeName(AuditTrail.timestampSortKeyName)
-          .attributeType(N)
-          .build(),
-        AttributeDefinition
-          .builder()
-          .attributeName(AuditTrail.userNameAttrName)
-          .attributeType(S)
-          .build()
-      )
-      .globalSecondaryIndexes(
-        GlobalSecondaryIndex
-          .builder()
-          .indexName(AuditTrail.secondaryIndexName)
-          .keySchema(
-            KeySchemaElement
-              .builder()
-              .attributeName(AuditTrail.userNameAttrName)
-              .keyType(HASH)
-              .build(),
-            KeySchemaElement
-              .builder()
-              .attributeName(AuditTrail.timestampSortKeyName)
-              .keyType(RANGE)
-              .build()
-          )
-          .projection(Projection.builder().projectionType("ALL").build())
-          .provisionedThroughput(
-            ProvisionedThroughput
-              .builder()
-              .readCapacityUnits(15L)
-              .writeCapacityUnits(15L)
-              .build()
-          )
-          .build()
-      )
-      .provisionedThroughput(
-        ProvisionedThroughput
-          .builder()
-          .readCapacityUnits(15L)
-          .writeCapacityUnits(15L)
-          .build()
-      )
-      .build()
-
-    dynamoDB.createTable(auditTableCreateRequest)
-  }
-
-  private[aws] def destroyTable()(implicit
-      dynamoDB: DynamoDbClient
-  ): DeleteTableResponse = {
-    val request =
-      DeleteTableRequest.builder().tableName(AuditTrail.tableName).build()
-    dynamoDB.deleteTable(request)
-  }
-}
-
-object AuditTrailDBTestApp {
-
-  def main(args: Array[String]): Unit = {
-    args.toList match {
-      case "create" :: Nil =>
-        new AuditTrailDBTest().createTable()(Clients.localDb)
-      case "destroy" :: Nil =>
-        new AuditTrailDBTest().destroyTable()(Clients.localDb)
-      case _ =>
-    }
   }
 
 }
