+++ test/aws/PasskeyChallengeDBTest.scala
@@ -72,84 +72,5 @@ class PasskeyChallengeDBTest extends AnyFreeSpec with Matchers {
       }
     }
 
-    "create table" ignore {
-      createTable()
-    }
-
-    "destroy table" ignore {
-      destroyTable()
-    }
   }
-
-  /** NB: Only use these for local testing use the provided CloudFormation
-    * template to create table in AWS environments.
-    *
-    * If you update this then be sure to also update the CloudFormation
-    * template's definition.
-    */
-  private[aws] def createTable()(implicit
-      dynamoDB: DynamoDbClient
-  ): CreateTableResponse = {
-    dynamoDB.createTable(
-      CreateTableRequest
-        .builder()
-        .tableName(tableName)
-        .keySchema(
-          KeySchemaElement
-            .builder()
-            .attributeName("username")
-            .keyType(HASH)
-            .build(),
-          KeySchemaElement
-            .builder()
-            .attributeName("flow")
-            .keyType(RANGE)
-            .build()
-        )
-        .attributeDefinitions(
-          AttributeDefinition
-            .builder()
-            .attributeName("username")
-            .attributeType(S)
-            .build(),
-          AttributeDefinition
-            .builder()
-            .attributeName("flow")
-            .attributeType(S)
-            .build()
-        )
-        .provisionedThroughput(
-          ProvisionedThroughput
-            .builder()
-            .readCapacityUnits(15L)
-            .writeCapacityUnits(15L)
-            .build()
-        )
-        .build()
-    )
-  }
-
-  private[aws] def destroyTable()(implicit
-      dynamoDb: DynamoDbClient
-  ): DeleteTableResponse =
-    dynamoDb.deleteTable(
-      DeleteTableRequest
-        .builder()
-        .tableName(tableName)
-        .build()
-    )
-}
-
-object PasskeyChallengeDBTestApp {
-
-  def main(args: Array[String]): Unit = {
-    args.toList match {
-      case "create" :: Nil =>
-        new PasskeyChallengeDBTest().createTable()(Clients.localDb)
-      case "destroy" :: Nil =>
-        new PasskeyChallengeDBTest().destroyTable()(Clients.localDb)
-      case _ =>
-    }
-  }
-
 }
