+++ test/aws/PasskeyDBTest.scala
@@ -9,82 +9,7 @@ import software.amazon.awssdk.services.dynamodb.model._
 
 class PasskeyDBTest extends AnyFreeSpec with Matchers {
 
-  "test db stuff - use this to test DynamoDB stuff locally during development" - {
-    given dynamoDB: DynamoDbClient = Clients.localDb
-
-    "create table" ignore {
-      createTable()
-    }
-
-    "destroy table" ignore {
-      destroyTable()
-    }
-  }
-
-  /** NB: Only use these for local testing use the provided CloudFormation
-    * template to create table in AWS environments.
-    *
-    * If you update this then be sure to also update the CloudFormation
-    * template's definition.
+  /** Once the table create code was removed, this test class was empty. TODO:
+    * Add some tests.
     */
-  private[aws] def createTable()(implicit
-      dynamoDB: DynamoDbClient
-  ): CreateTableResponse =
-    dynamoDB.createTable(
-      CreateTableRequest
-        .builder()
-        .tableName(PasskeyDB.tableName)
-        .keySchema(
-          KeySchemaElement
-            .builder()
-            .attributeName("username")
-            .keyType(HASH)
-            .build(),
-          KeySchemaElement
-            .builder()
-            .attributeName("credentialId")
-            .keyType(RANGE)
-            .build()
-        )
-        .attributeDefinitions(
-          AttributeDefinition
-            .builder()
-            .attributeName("username")
-            .attributeType(S)
-            .build(),
-          AttributeDefinition
-            .builder()
-            .attributeName("credentialId")
-            .attributeType(S)
-            .build()
-        )
-        .provisionedThroughput(
-          ProvisionedThroughput
-            .builder()
-            .readCapacityUnits(15L)
-            .writeCapacityUnits(15L)
-            .build()
-        )
-        .build()
-    )
-
-  private[aws] def destroyTable()(implicit
-      dynamoDb: DynamoDbClient
-  ): DeleteTableResponse =
-    dynamoDb.deleteTable(
-      DeleteTableRequest.builder().tableName(PasskeyDB.tableName).build()
-    )
-}
-
-object PasskeyDBTestApp {
-
-  def main(args: Array[String]): Unit = {
-    args.toList match {
-      case "create" :: Nil => new PasskeyDBTest().createTable()(Clients.localDb)
-      case "destroy" :: Nil =>
-        new PasskeyDBTest().destroyTable()(Clients.localDb)
-      case _ =>
-    }
-  }
-
 }
